feat/#353 : 함께 패킹리스트 초대 모달에서 ssr 하는 방식으로 수정
